requirements.txt
python-dotenv
azure-identity
azure-storage-blob
pandas
pyarrow
